calling location api
